+++ src/realmrelay/game/WebMain.java
@@ -4,8 +4,8 @@ import realmrelay.game.util.AssetLoader;
 
 public class WebMain {
 
-    public static float sWidth = 800;
-    public static float sHeight = 600;
+    public static double sWidth = 800;
+    public static double sHeight = 600;
 
     public static void main(String[] args) {
         WebMain w = new WebMain();
